+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog
 =========
 
 ## 1.0.1 (2020-06-14)
+ * Added assets (gitignore fix)
  * Update README.md
 
 ## 1.0.0 (2020-06-07)